move type tag to types.rs
